More small tweaks on server library

+++ src/nrf24l01_proto_net.h
@@ -65,7 +65,7 @@ typedef struct __attribute__ ((packed)) {
 	uint8_t				maj_version;
 	uint8_t				min_version;
 	uint32_t				hashid;
-	uint8_t				data;
+	uint32_t				data;
 	int8_t					result;
 } nrf24_join_local;
